pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto_generator"
version = "0.1.0"
dependencies = [
    "bs4",
    "Flask",
    "html5validator",
    "pycodestyle",
    "pydocstyle",
    "pylint",
    "pytest",
    "requests",
    "pytest-mock",
    "fpdf2",
]
requires-python = ">=3.8"

[tool.setuptools]
packages = ["auto_generator"]